Fix - Reduce RAM to fit instance
